backend/requirements.txt
# AI Assistant Backend Dependencies

# Web Framework
fastapi==0.115.6
uvicorn==0.32.1

# Database
supabase==2.10.0

# AI / LLM
anthropic==0.40.0
openai==1.57.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.27.2


python-multipart==0.0.20